feat: add validation to avoid deletion of users and vehicles with active bookings

+++ src/modules/user/user.service.ts
@@ -5,8 +5,16 @@ const getUsers = async () => {
   return result;
 };
 
-// TODO: Delete a user (only if no active bookings exist)
 const deleteUser = async (id: string) => {
+  const bookingResult = await pool.query(
+    `SELECT customer_id FROM bookings WHERE customer_id=$1`,
+    [id]
+  );
+
+  if (bookingResult.rowCount && bookingResult.rowCount > 0) {
+    throw new Error("Can't delete user with active bookings");
+  }
+
   const result = await pool.query(`DELETE FROM users WHERE id = $1`, [id]);
   return result;
 };
